+++ src/test/kotlin/one/wabbit/typeclass/plugin/integration/ImportVisibilityTest.kt
@@ -216,6 +216,244 @@ class ImportVisibilityTest : IntegrationTestSupport() {
         )
     }
 
+    @Test fun topLevelInstancesMayBeDeclaredBesideNestedDeclaredHeadArguments() {
+        val sources =
+            mapOf(
+                "shared/Api.kt" to
+                    """
+                    package shared
+
+                    import one.wabbit.typeclass.Typeclass
+
+                    @Typeclass
+                    interface Rel<A, B> {
+                        fun label(): String
+                    }
+
+                    context(rel: Rel<A, B>)
+                    fun <A, B> which(): String = rel.label()
+                    """.trimIndent(),
+                "alpha/Foo.kt" to
+                    """
+                    package alpha
+
+                    data class Foo(val value: Int = 0)
+                    """.trimIndent(),
+                "beta/Boo.kt" to
+                    """
+                    package beta
+
+                    data class Boo<A>(val value: A)
+                    """.trimIndent(),
+                "gamma/Baz.kt" to
+                    """
+                    package gamma
+
+                    import alpha.Foo
+                    import beta.Boo
+                    import one.wabbit.typeclass.Instance
+                    import shared.Rel
+
+                    data class Baz(val value: Int)
+
+                    @Instance
+                    object FooBooBazRel : Rel<Foo, Boo<Baz>> {
+                        override fun label(): String = "baz-host"
+                    }
+                    """.trimIndent(),
+                "demo/Main.kt" to
+                    """
+                    package demo
+
+                    import alpha.Foo
+                    import beta.Boo
+                    import gamma.Baz
+                    import shared.which
+
+                    fun main() {
+                        println(which<Foo, Boo<Baz>>())
+                    }
+                    """.trimIndent(),
+            )
+
+        assertCompilesAndRuns(
+            sources = sources,
+            expectedStdout = "baz-host",
+            mainClass = "demo.MainKt",
+        )
+    }
+
+    @Test fun topLevelInstancesMayBeDeclaredBesideOuterDeclaredHeadArguments() {
+        val sources =
+            mapOf(
+                "shared/Api.kt" to
+                    """
+                    package shared
+
+                    import one.wabbit.typeclass.Typeclass
+
+                    @Typeclass
+                    interface Rel<A, B> {
+                        fun label(): String
+                    }
+
+                    context(rel: Rel<A, B>)
+                    fun <A, B> which(): String = rel.label()
+                    """.trimIndent(),
+                "alpha/Foo.kt" to
+                    """
+                    package alpha
+
+                    data class Foo(val value: Int = 0)
+                    """.trimIndent(),
+                "beta/Boo.kt" to
+                    """
+                    package beta
+
+                    import alpha.Foo
+                    import gamma.Baz
+                    import one.wabbit.typeclass.Instance
+                    import shared.Rel
+
+                    data class Boo<A>(val value: A)
+
+                    @Instance
+                    object FooBooBazRel : Rel<Foo, Boo<Baz>> {
+                        override fun label(): String = "boo-host"
+                    }
+                    """.trimIndent(),
+                "gamma/Baz.kt" to
+                    """
+                    package gamma
+
+                    data class Baz(val value: Int)
+                    """.trimIndent(),
+                "demo/Main.kt" to
+                    """
+                    package demo
+
+                    import alpha.Foo
+                    import beta.Boo
+                    import gamma.Baz
+                    import shared.which
+
+                    fun main() {
+                        println(which<Foo, Boo<Baz>>())
+                    }
+                    """.trimIndent(),
+            )
+
+        assertCompilesAndRuns(
+            sources = sources,
+            expectedStdout = "boo-host",
+            mainClass = "demo.MainKt",
+        )
+    }
+
+    @Test fun rejectsTopLevelInstancesDeclaredOutsideDeclaredHeadOwnersEvenForNestedArguments() {
+        val sources =
+            mapOf(
+                "shared/Api.kt" to
+                    """
+                    package shared
+
+                    import one.wabbit.typeclass.Typeclass
+
+                    @Typeclass
+                    interface Rel<A, B> {
+                        fun label(): String
+                    }
+                    """.trimIndent(),
+                "alpha/Foo.kt" to
+                    """
+                    package alpha
+
+                    data class Foo(val value: Int = 0)
+                    """.trimIndent(),
+                "beta/Boo.kt" to
+                    """
+                    package beta
+
+                    data class Boo<A>(val value: A)
+                    """.trimIndent(),
+                "gamma/Baz.kt" to
+                    """
+                    package gamma
+
+                    data class Baz(val value: Int)
+                    """.trimIndent(),
+                "qux/Qux.kt" to
+                    """
+                    package qux
+
+                    import alpha.Foo
+                    import beta.Boo
+                    import gamma.Baz
+                    import one.wabbit.typeclass.Instance
+                    import shared.Rel
+
+                    data class Qux(val value: Int = 0)
+
+                    @Instance
+                    object FooBooBazRel : Rel<Foo, Boo<Baz>> { // E:TC_INVALID_INSTANCE_DECL nested declared heads should not make unrelated files legal hosts
+                        override fun label(): String = "qux-host"
+                    }
+                    """.trimIndent(),
+            )
+
+        assertDoesNotCompile(
+            sources = sources,
+            expectedDiagnostics = listOf(expectedInvalidInstanceDecl("same file", "rel", "foo", "boo", "baz")),
+        )
+    }
+
+    @Test fun entailedSupertypesDoNotConferTopLevelOrphanOwnership() {
+        val sources =
+            mapOf(
+                "shared/Eq.kt" to
+                    """
+                    package shared
+
+                    import model.Foo
+                    import one.wabbit.typeclass.Instance
+                    import one.wabbit.typeclass.Typeclass
+
+                    @Typeclass
+                    interface Eq<A> {
+                        fun eqv(left: A, right: A): Boolean
+                    }
+
+                    @Instance
+                    object FooOrd : Ord<Foo> { // E:TC_INVALID_INSTANCE_DECL declared Ord<Foo> should not become legal in Eq.kt via entailed Eq<Foo>
+                        override fun eqv(left: Foo, right: Foo): Boolean = left == right
+                        override fun compare(left: Foo, right: Foo): Int = left.value.compareTo(right.value)
+                    }
+                    """.trimIndent(),
+                "shared/Ord.kt" to
+                    """
+                    package shared
+
+                    import one.wabbit.typeclass.Typeclass
+
+                    @Typeclass
+                    interface Ord<A> : Eq<A> {
+                        fun compare(left: A, right: A): Int
+                    }
+                    """.trimIndent(),
+                "model/Foo.kt" to
+                    """
+                    package model
+
+                    data class Foo(val value: Int)
+                    """.trimIndent(),
+            )
+
+        assertDoesNotCompile(
+            sources = sources,
+            expectedDiagnostics = listOf(expectedInvalidInstanceDecl("same file", "ord", "foo")),
+        )
+    }
+
     @Test fun internalDependencyInstancesDeclaredBesideTypeclassHeadsDoNotLeakAcrossModuleBoundaries() {
         val dependency =
             HarnessDependency(